Fixed segmentation fault
The __nxml_get_value function does not check the memory boundaries, and
reads beyond the end of the buffer.  This reads bogus if the xml comes
from a buffer, and goes segfault if the xml comes from a file.

This patch also includes a test case that highlights the issue.