Clarified contract with __nxml_parse_get_attr
As I understand, this function should extract the next attribute of a tag.
It is difficult to understand, as it relies on macros that change the
inner state.  Really bad practices.

The function is expected to return NULL if no attribute is available, and
an allocated string otherwise.  In the original implementation it is
however impossible to distinguish the case of a faulty memory allocation
by strdup.

The modified version of __nxml_parse_get_attr has a clear contract: the
return value determines what happens, and the *result buffer is assigned
with the allocated string.

The modification also prevents a suspected buffer overflow that might
happen if 'i' is incremented beyond the size of  the 'attr' buffer.  It is
not clear if this can actually happen, because of all those filthy macros,
but better safe than sorry.

The only place where __nxml_parse_get_attr is invoked is updated
accordingly.