Fix more crashes
This patch fixes a crash affecting the library under OpenBSD and FreeBSD
(and possibly more systems).  Interestingly, nothing happens under
GNU/Linux.

The conditionals should first check the available buffer size, and only
then then access the buffer, even if it is just read.  This is because the
C programming language has short-circuit conditionals.  In other places,
the buffer is just accessed blindly.

I'm not sure why it breaks only under BSD.  My educated guess: it might
have to do with memory layout (e.g. reading a memory location that is not
mapped at all in the process' memory).

In general, checking the buffer size is always important, yet it is often
neglected in this codebase.  I would recommend to review the code, and
abstract away the boundary checking, so that it happens systematically.
This patch only focuses on solving an individual occurrence, triggered by
a specific XML file.